requirements.txt
# Core ML / Data
numpy==2.3.4
pandas==2.3.3
scikit-learn==1.7.2
xgboost==3.1.1
joblib==1.5.2

# Visualization
matplotlib==3.10.7
seaborn==0.13.2
plotly==6.3.1

# Configuration & Utilities
PyYAML==6.0.3

# Experiment tracking & data versioning
mlflow==3.5.1
dagshub==0.6.3
dvc-s3==3.2.2

# Testing / notebooks
pytest==8.4.2
jupyter==1.1.1
